setup auto update skin
This should work. On the 1st of March, June, Sep, and Dec, the update season skin action will be triggered. This will read the current season skin and use script/advance_season.sh to advance a season, then change the content of _config.yml and commit and push to the master branch. Be aware of temporary writing permission granted to github bot and potential security concerns.